backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<4.5
django-environ>=0.11,<0.12
psycopg2-binary>=2.9,<2.10
Pillow>=10.2,<10.4
opencv-python-headless>=4.13,<4.14
python-multipart>=0.0,<0.1
gunicorn>=21.2,<21.3
drf-spectacular>=0.27,<0.29
fastapi>=0.111,<0.116
SQLAlchemy>=2.0,<2.1
uvicorn>=0.30,<0.31
supabase>=2.4,<3.0
